Cambios.
Cambio de nombre: obraarte =>  ObraArte 
Eliminados atributos: titulo, dimenciones
Agregado atributo: caracteristica 
Renombrado: dimenciones => valor 
Método mostrar() ahora es abstracto
Constructores actualizados con nuevos parámetros
